fmq_test_client device -> general-tests
We don't need to build this per-device.

This didn't work w/ general-tests before because tradefed was running
32-bit tests from test_suites_x86_64 on 32-bit devices, which caused
this test to fail, since it still installed 64-bit tests in that case,
and this test tried to run it. Although a separate bug is filed to
investigate that, switching the method of abi detection here.

Bug: N/A
Test: fmq_test
Change-Id: Ief95ba52255300ee990b8c9d9fd585b194c8aff5