fmq_fuzzer: add blocking read and write calls
Test the Synchronous blocking read and write APIs.
Spawn a reader thread that will keep reading until it times out on a
blocking read. This will mostly happen when the writer has exhausted its
writes, but will also happen if the writer is blocking on a large write
that won't yet fit in the queue.
The writer on the main thread will continue writing as much as it can.

Test: m fmq_fuzzer &&
./out/host/linux-x86/fuzz/x86_64/fmq_fuzzer/fmq_fuzzer
Bug: 165621851

Change-Id: I122ddba837f3ce995e720182dcd6abc8cdb71701