fmq_fuzzer: Exit blocking threads when no more work
One configuration of the fuzzer will run a blocking writer thread with a
number of blocking reader threads. The blocking operations time out
after 10ms. It's possible for one side (readers or writer) to end up
with thousands of extra operations to do after the other side has
finished their job. This leads to thousands of timeouts, greatly
increasing the time of the test.
Since there is no benifit to coniously time out, we now exit when that
happens.

Bug: 218521670
Test: Download corpus from failing runs
Test: m SANITIZE_HOST=address fmq_fuzzer
Test: ./fmq_fuzzer -max_len=50000 -timeout=25 -rss_limit_mb=2560 -use_value_profile=1 -max_total_time=2700 -print_final_stats=1 -artifact_prefix=/usr/local/google/home/devinmoore/workspace/fmq_fuzzer_artifacts  corpus/
Change-Id: Icfe014f07b9838ac28a55cf3f010ee8f33a346de